Added login/signup option

+++ app/models/Model_posts.php
@@ -7,28 +7,35 @@ use app\core\Model;
 
 class Model_posts extends Model
 {
-    public function getAllPosts($order_by = null)
+    public function getAllPosts($order_by=null)
     {
         $dsn = "$this->db_driver:host=$this->db_host; dbname=$this->db_name; charset=$this->db_charset";
         $user = $this->db_username;
         $password = $this->db_password;
 
-        switch ($order_by){
-            case 'date_asc':
-                $sql = 'SELECT * FROM posts ORDER BY created_at ASC';
-                break;
-            case 'date_desc':
-                $sql = 'SELECT * FROM posts ORDER BY created_at DESC';
-                break;
-            case 'header_asc':
-                $sql = 'SELECT * FROM posts ORDER BY header ASC';
-                break;
-            case 'header_desc':
-                $sql = 'SELECT * FROM posts ORDER BY header DESC';
-                break;
-            default :
-                $sql = 'SELECT * FROM posts';
-        }
+            switch ($order_by) {
+                case 'date_asc':
+                    //$sql = 'SELECT * FROM posts ORDER BY created_at ASC';
+                    $sql = 'SELECT posts.id, posts.header, posts.text, posts.created_at, posts.edited_at, users.login 
+                        FROM posts INNER JOIN users ON posts.author_id=users.id';
+                    break;
+                case 'date_desc':
+                    $sql = 'SELECT posts.id, posts.header, posts.text, posts.created_at, posts.edited_at, users.login 
+                        FROM posts INNER JOIN users ON posts.author_id=users.id ORDER BY posts.created_at DESC';
+                    break;
+                case 'header_asc':
+                    $sql = 'SELECT posts.id, posts.header, posts.text, posts.created_at, posts.edited_at, users.login 
+                        FROM posts INNER JOIN users ON posts.author_id=users.id ORDER BY posts.header ASC';
+                    break;
+                case 'header_desc':
+                    $sql = 'SELECT posts.id, posts.header, posts.text, posts.created_at, posts.edited_at, users.login 
+                        FROM posts INNER JOIN users ON posts.author_id=users.id ORDER BY posts.header DESC';
+                    break;
+                default :
+                    //$sql = 'SELECT * FROM posts';
+                    $sql = 'SELECT posts.id, posts.header, posts.text, posts.created_at, posts.edited_at, users.login 
+                        FROM posts INNER JOIN users ON posts.author_id=users.id';
+            }
 
         try {
             $connection = new \PDO($dsn,$user,$password);
@@ -42,11 +49,50 @@ class Model_posts extends Model
         return $posts_array;
     }
 
+    public function getAllPostsByUser($user_id, $order_by=null)
+    {
+        $dsn = "$this->db_driver:host=$this->db_host; dbname=$this->db_name; charset=$this->db_charset";
+        $user = $this->db_username;
+        $password = $this->db_password;
+
+        switch ($order_by){
+            case 'date_asc':
+                $sql = 'SELECT * FROM posts WHERE author_id = ? ORDER BY created_at ASC';
+                break;
+            case 'date_desc':
+                $sql = 'SELECT * FROM posts WHERE author_id = ? ORDER BY created_at DESC';
+                break;
+            case 'header_asc':
+                $sql = 'SELECT * FROM posts WHERE author_id = ? ORDER BY header ASC';
+                break;
+            case 'header_desc':
+                $sql = 'SELECT * FROM posts WHERE author_id = ? ORDER BY header DESC';
+                break;
+            default :
+                $sql = 'SELECT * FROM posts WHERE author_id = ? ORDER BY created_at ASC';
+                break;
+        }
+
+        try {
+            $connection = new \PDO($dsn,$user,$password);
+            $stmt = $connection->prepare($sql);
+            $stmt->execute([$user_id]);
+            $posts_array = $stmt->fetchAll(\PDO::FETCH_ASSOC);
+
+
+        } catch (\PDOException $e) {
+            return $e->getMessage();
+        }
+
+        return $posts_array;
+
+    }
+
     /**
      * @param $id
      * @return array . One post by @param $id.
      */
-    public function getPost($id)
+    public function getPost($post_id, $user_id, $role_super_admin=null)
     {
         $dsn = "$this->db_driver:host=$this->db_host; dbname=$this->db_name; charset=$this->db_charset";
         $user = $this->db_username;
@@ -54,10 +100,24 @@ class Model_posts extends Model
 
         try {
             $connection = new \PDO($dsn,$user,$password);
-            $sql = "SELECT * FROM posts WHERE id = ?";
-            $stmt = $connection->prepare($sql);
-            $stmt->execute([$id]);
-            $post_array = $stmt->fetch(\PDO::FETCH_ASSOC);
+            if ($role_super_admin){
+                $sql = "SELECT * FROM posts WHERE id = ?";
+                $stmt = $connection->prepare($sql);
+                $stmt->execute([$post_id]);
+                $post_array = $stmt->fetch(\PDO::FETCH_ASSOC);
+                if (empty($post_array)) {
+                    return false;
+                }
+            } else {
+
+                $sql = "SELECT * FROM posts WHERE id = ? AND author_id = ?";
+                $stmt = $connection->prepare($sql);
+                $stmt->execute([$post_id, $user_id]);
+                $post_array = $stmt->fetch(\PDO::FETCH_ASSOC);
+                if (empty($post_array)) {
+                    return false;
+                }
+            }
 
         } catch (\PDOException $e) {
             return $e->getMessage();
@@ -75,6 +135,7 @@ class Model_posts extends Model
     {
         $header = $data['header'];
         $text = $data['text'];
+        $author_id = $data['author_id'];
 
         $dsn = "$this->db_driver:host=$this->db_host; dbname=$this->db_name; charset=$this->db_charset";
         $user = $this->db_username;
@@ -83,10 +144,11 @@ class Model_posts extends Model
 
         try {
             $connection = new \PDO($dsn,$user,$password);
-            $sql = "INSERT INTO posts VALUES (null,:header, :text, now(), NULL)";
+            $sql = "INSERT INTO posts VALUES (null,:header, :text, now(), NULL, :author_id)";
             $stmt = $connection->prepare($sql);
-            $stmt->execute(['header'=>$header,
-                            'text'=>$text]);
+            $stmt->execute(['header'    =>$header,
+                            'text'      =>$text,
+                            'author_id' =>$author_id]);
 
             return true;
 
@@ -99,7 +161,7 @@ class Model_posts extends Model
      * @param $id
      * @return bool
      */
-    public function destroy($id)
+    public function destroy($post_id, $user_id, $role_super_admin=null)
     {
         $dsn = "$this->db_driver:host=$this->db_host; dbname=$this->db_name; charset=$this->db_charset";
         $user = $this->db_username;
@@ -107,13 +169,27 @@ class Model_posts extends Model
 
         try {
             $connection = new \PDO($dsn,$user,$password);
-            $sql = "DELETE FROM posts WHERE id = ?";
-            $stmt = $connection->prepare($sql);
-            $stmt->execute([$id]);
+            if ($role_super_admin) {
 
+                $sql = "DELETE FROM posts WHERE id = ?";
+                $stmt = $connection->prepare($sql);
+                $stmt->execute([$post_id]);
+                $result = $stmt->rowCount();
+                if ($result == 0) {
+                    return false;
+                }
+            } else {
+                $sql = "DELETE FROM posts WHERE id = ? AND author_id = ?";
+                $stmt = $connection->prepare($sql);
+                $stmt->execute([$post_id, $user_id]);
+                $result = $stmt->rowCount();
+                if ($result == 0) {
+                    return false;
+                }
+            }
             return true;
 
-        } catch (\PDOException $e) {
+        } catch (\Exception $e) {
             return $e->getMessage();
         }
     }
@@ -123,9 +199,9 @@ class Model_posts extends Model
      * @param $post
      * @return bool
      */
-    public function edit($post)
+    public function edit($post, $user_id, $role_super_admin=false)
     {
-        $id = $post['id'];
+        $post_id = $post['id'];
         $header = $post['header'];
         $text = $post['text'];
 
@@ -136,9 +212,28 @@ class Model_posts extends Model
 
         try {
             $connection = new \PDO($dsn,$user,$password);
-            $sql = "UPDATE posts SET header = :header, text = :text, edited_at = now() WHERE id = :id";
-            $stmt = $connection->prepare($sql);
-            $stmt->execute(['header'=>$header, 'text'=>$text, 'id'=>$id]);
+            if ($role_super_admin) {
+                $sql = "UPDATE posts SET header = :header, text = :text, edited_at = now() WHERE id = :post_id";
+                $stmt = $connection->prepare($sql);
+                $stmt->execute(['header'=>$header,
+                                'text'=>$text,
+                                'post_id'=>$post_id]);
+                $result = $stmt->rowCount();
+                if ($result == 0) {
+                    return false;
+                }
+            } else {
+                $sql = "UPDATE posts SET header = :header, text = :text, edited_at = now() WHERE id = :post_id AND author_id= :author_id";
+                $stmt = $connection->prepare($sql);
+                $stmt->execute(['header'=>$header,
+                    'text'=>$text,
+                    'post_id'=>$post_id,
+                    'author_id'=>$user_id]);
+                $result = $stmt->rowCount();
+                if ($result == 0) {
+                    return false;
+                }
+            }
 
             return true;
